Add Vulnerabilities and license checks
Creates two checks:

- A check with vulnerabilities added. If the check fails, it adds to the
check the list of vulnerabilities found grouped by manifest. The check
fails if vulnerables packages are added.
- License check. A license check is created and it lists on the detail
incompatible licenses (if any) and the unknown licenses. The check fails
if there are incompatible licenses and succeeds otherwise

Checks names are configurable.

A breaking change was introduced, the action no longer fails if there are
vulnerabilities or license violations. fail-on-violation parameter
can be set to true to maintain backward compatibility.